ceo_assistant/requirements.txt
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-api-python-client>=2.111.0
google-auth-oauthlib>=1.2.0
